pyproject.toml
[project]
name = "dj-bita-backend"
version = "0.1.0"
description = "Bita Suits API services backend"
authors = [
    {name = "Murad",email = "nuradhussen082@gmail.com"}
]
license = {text = "Apache 2.0"}
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "django-guardian (>=3.0.3,<4.0.0)",
    "django (>=5.2.4,<6.0.0)",
    "django-restframework (>=0.0.1,<0.0.2)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "daphne (>=4.2.1,<5.0.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)",
    "django-storages[s3] (>=1.14.6,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "channels[daphne] (>=4.2.2,<5.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "boto3 (>=1.39.4,<2.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "django-admin-interface (>=0.30.1,<0.31.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "firebase-admin (>=7.1.0,<8.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
zappa = "^0.60.2"
black = "^25.1.0"
isort = "^6.0.1"
